requirements.txt
streamlit>=1.28.0
databricks-sdk>=0.12.0
requests>=2.31.0
PyPDF2>=3.0.1
python-dotenv>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
python-multipart>=0.0.6
openai>=1.0.0
tiktoken>=0.5.0
